move model functions to own file
